fix(ios): inline theme types into ContentView (Theme.swift was not in the project)
Theme.swift was never added to the XcodeGen-generated project, so its
symbols (AppAppearance, appAppearanceStorageKey, Color.ci*) weren't compiled
— breaking the build. Move them into ContentView.swift (already in the
project) and drop the standalone file, avoiding an xcodegen regen that would
clobber local bundle-id/signing changes in the pbxproj.
